Re-organize main test program model
