+++ rust/origen/src/prog_gen/teradyne/base/mod.rs
@@ -1,4 +1,4 @@
-use super::super::tests::ParamType;
+use super::super::model::ParamType;
 use super::super::TestProgram;
 use crate::Result;
